Move files to migration-graph-ember; bump swc/core
