+++ packages/migration-graph-ember/package.json
@@ -22,6 +22,7 @@
     "@babel/parser": "^7.18.13",
     "@babel/traverse": "^7.18.13",
     "@rehearsal/migration-graph-shared": "^0.0.37",
+    "@swc/core": "^1.3.2",
     "fast-glob": "^3.2.12",
     "find-up": "^5.0.0",
     "fs-extra": "^10.1.0",
